+++ LicenseManagement.EndUser/LicenseHandlingUninstall.cs
@@ -17,9 +17,19 @@ namespace LicenseManagement.EndUser
 
         void SetNextHandler()
         {
-
-            //we must have internet so we can un
-            NextHandler = new ApiGetComputerHandler();
+            // Read the existing on-disk license to get its ID for the server-side PATCH.
+            // Unlike install/launch, uninstall does not need to re-register the computer.
+            var register = new Registrars.LicenseRegister(HandlingContext);
+            if (register.TryRead(out string signedLic))
+            {
+                HandlingContext.SetLicenseData(signedLic, false);
+                NextHandler = new Signature.Handlers.LicenseSignatureValidationHandler();
+            }
+            else
+            {
+                HandlingContext.Exception = new Exceptions.CouldNotReadLicenseFromDiskException();
+                NextHandler = new ErrorHandler();
+            }
         }
         public override void HandleLicense()
         {